+++ CHANGELOG.md
@@ -3,16 +3,19 @@
 ## UNRELEASED - Runtime
 
 - fix: doc comments from object fields should be present in generated GraphQL schema [#630](https://github.com/hypermodeinc/modus/pull/630)
+- feat: add neo4j support in modus [#636](https://github.com/hypermodeinc/modus/pull/636)
 
 ## UNRELEASED - Go SDK
 
 - fix: vector package should return generic type in computations [#628](https://github.com/hypermodeinc/modus/pull/628)
 - chore: Remove unused go package reference [#632](https://github.com/hypermodeinc/modus/pull/632)
+- feat: add neo4j support in modus [#636](https://github.com/hypermodeinc/modus/pull/636)
 
 ## UNRELEASED - AssemblyScript SDK
 
 - fix: vector package should return generic type in computations [#628](https://github.com/hypermodeinc/modus/pull/628)
 - chore: Delete extraneous copy of Anthropic model interface [#631](https://github.com/hypermodeinc/modus/pull/631)
+- feat: add neo4j support in modus [#636](https://github.com/hypermodeinc/modus/pull/636)
 - feat: Add `DynamicMap` type [#638](https://github.com/hypermodeinc/modus/pull/638)
 
 ## 2024-11-27 - CLI 0.14.0